# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "askPMO API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "description": "Liveness: the process is up.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Ready",
        "description": "Readiness: index, database, queue and Foundry reachable; 503 otherwise.",
        "operationId": "ready_api_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyOut"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "description": "List profiles with their resolved toggles.",
        "operationId": "list_profiles_api_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileOut"
                  },
                  "type": "array",
                  "title": "Response List Profiles Api Profiles Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Documents",
        "description": "All documents ordered by relative path.",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "type": "array",
                  "title": "Response List Documents Api Documents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Documents",
        "description": "Store files, create one ingest job and enqueue it; unchanged re-uploads are skipped.",
        "operationId": "upload_documents_api_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_api_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest-jobs/{job_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Ingest Job",
        "description": "Progress of one ingest job.",
        "operationId": "get_ingest_job_api_ingest_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chunks/{chunk_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Chunk",
        "description": "Full chunk text and metadata, for a clicked citation.",
        "operationId": "get_chunk_api_chunks__chunk_id__get",
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_documents_api_documents_post": {
        "properties": {
          "files[]": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files[]"
          },
          "rel_paths[]": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rel Paths[]"
          }
        },
        "type": "object",
        "required": [
          "files[]",
          "rel_paths[]"
        ],
        "title": "Body_upload_documents_api_documents_post"
      },
      "ChunkOut": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "doc_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Date"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          },
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "text",
          "source",
          "section",
          "doc_type",
          "doc_date",
          "week",
          "ids"
        ],
        "title": "ChunkOut",
        "description": "Full chunk text and metadata, for a clicked citation."
      },
      "DocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "rel_path": {
            "type": "string",
            "title": "Rel Path"
          },
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "doc_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Date"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "indexed",
              "failed"
            ],
            "title": "Status"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "rel_path",
          "doc_type",
          "doc_date",
          "week",
          "status",
          "chunk_count",
          "created_at"
        ],
        "title": "DocumentOut",
        "description": "A stored document."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthOut": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthOut",
        "description": "Liveness."
      },
      "IngestJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "document_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Document Ids"
          },
          "documents_done": {
            "type": "integer",
            "title": "Documents Done"
          },
          "chunks_written": {
            "type": "integer",
            "title": "Chunks Written"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "document_ids",
          "documents_done",
          "chunks_written",
          "error",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "IngestJobOut",
        "description": "Ingest job progress."
      },
      "ProfileOut": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "baseline",
              "hybrid",
              "rerank",
              "full"
            ],
            "title": "Name"
          },
          "toggles": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Toggles"
          }
        },
        "type": "object",
        "required": [
          "name",
          "toggles"
        ],
        "title": "ProfileOut",
        "description": "A profile and its resolved toggles."
      },
      "ReadyCheck": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "title": "ReadyCheck",
        "description": "Result of one dependency check."
      },
      "ReadyOut": {
        "properties": {
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "checks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReadyCheck"
            },
            "type": "object",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "ready",
          "checks"
        ],
        "title": "ReadyOut",
        "description": "Readiness: index, database, queue and Foundry."
      },
      "UploadResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "document_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Document Ids"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "document_ids"
        ],
        "title": "UploadResponse",
        "description": "Result of an upload: one job for all accepted documents."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}